\name{aracne2regulon}
\alias{aracne2regulon}
\title{Regulon object generation from ARACNe results}
\usage{
  aracne2regulon(afile, eset, gene = F,
    format = c("adj", "3col")[1], verbose = T)
}
\arguments{
  \item{afile}{Character string indicating the name of the
  ARACNe network file}

  \item{eset}{Either a character string indicating the name
  of the expression-dataset file, or a gene expression
  matrix with genes (probes) in rows and samples in
  columns}

  \item{gene}{Logical, whether the probes should be
  collapsed at the gene level}

  \item{format}{Character string, indicating the format of
  the aracne file, either adj for adjacency matrixes
  generated by aracne, or 3col when the interactome is
  represented by a 3 columns text file, with regulator in
  the first column, target in the second and mutual
  information in the third column}

  \item{verbose}{Logical, whether progression messages
  should be printed in the terminal.}
}
\value{
  Regulon object
}
\description{
  This function generates a regulon object from ARACNe
  results and the corresponding expression dataset
}
\seealso{
  \code{\link{marina}}, \code{\link{ssmarina}}
}

