\name{bootstrapMarina}
\alias{bootstrapMarina}
\title{MARINA bootstraps integration}
\usage{
  bootstrapMarina(mobj, method = "mode")
}
\arguments{
  \item{mobj}{MARINA object}

  \item{method}{Character string indicating the method to
  use, either mean, median or mode}
}
\value{
  MARINA object
}
\description{
  This function integrates the bootstrap MARINA results
}
\seealso{
  \code{\link{marina}}
}

