\name{comNames}
\alias{comNames}
\title{Combinatorial annotation}
\usage{
  comNames(x, annot)
}
\arguments{
  \item{x}{Character vector of gene name combinations,
  where the combinations are separated by --}

  \item{annot}{Vector of gene names with geneID as
  \code{names} attribute}
}
\value{
  Converted annotations
}
\description{
  This function convers combinatorial annotations
}
\seealso{
  \code{\link{marina}}
}

