\name{distMode}
\alias{distMode}
\title{Mode of continuous distributions}
\usage{
  distMode(x, adj = 1)
}
\arguments{
  \item{x}{Numeric data vector}

  \item{adj}{Number indicating the adjustment for the
  kernel bandwidth}
}
\value{
  Number
}
\description{
  This function computes the mode for continuous
  distributions
}

