\name{fclengthna}
\alias{fclengthna}
\title{Length of columns of arrays with NAs}
\usage{
  fclengthna(x)
}
\arguments{
  \item{x}{Matrix}
}
\value{
  1-column matrix of integers indicating the number of
  non-NA columns
}
\description{
  This function report the number of columns of a matrix
  ignoring NA values
}

