\name{fcvarna}
\alias{fcvarna}
\title{Variance of columns for arrays with NA values}
\usage{
  fcvarna(x)
}
\arguments{
  \item{x}{Numeric matrix}
}
\value{
  1-column matrix with the variance by column results
}
\description{
  This function computes the variance by columns ignoring
  NA values
}

