\name{frlengthna}
\alias{frlengthna}
\title{Length of rows for arrays with NA values}
\usage{
  frlengthna(x)
}
\arguments{
  \item{x}{Matrix}
}
\value{
  1-column matrix of integers indicating the number of
  non-NA rows
}
\description{
  This function report the length of rows of a matrix
  ignoring the NA values
}

