\name{ledge}
\alias{ledge}
\title{Leading-edge analysis}
\usage{
  ledge(mobj)
}
\arguments{
  \item{mobj}{marina class object}
}
\value{
  marina object updated with a ledge slot
}
\description{
  This function performs a Leading-Edge analysis on an
  object of class marina
}
\seealso{
  \code{\link{marina}}
}

