\name{marinaAnnot}
\alias{marinaAnnot}
\title{MARINa annotation change}
\usage{
  marinaAnnot(mobj, annot, complete = T)
}
\arguments{
  \item{mobj}{MARINa object generated by \code{marina}
  function}

  \item{annot}{Vector os character strings containing the
  gene names and gene identifiers as vector names
  attribute}

  \item{complete}{Logical, whether the signature and target
  names should be also transformed}
}
\value{
  MARINa object with updated annotations
}
\description{
  This function changes the annotation of genes in MARINa
  objects
}
\seealso{
  \code{\link{marina}}
}

