\name{marinaCombinatorial}
\alias{marinaCombinatorial}
\title{MARINA combinatorial analysis}
\usage{
  marinaCombinatorial(mobj, regulators = 100,
    nullmodel = NULL, minsize = NULL, adaptive.size = NULL,
    iterative = NULL, level = 10, verbose = T)
}
\arguments{
  \item{mobj}{MARINa object generated by \code{marina}
  function}

  \item{regulators}{Either a number between 0 and 1
  indicating the p-value cutoff for individual TFs to be
  included in the combinations analysis; (>1) indicating
  the number of top TFs to be included in the combinations
  analysis; or a vector of character strings indicating the
  TF IDs to be included in the analysis}

  \item{nullmodel}{Matrix of genes by permutations
  containing the NULL model signatures. Taken from
  \code{mobj} by default}

  \item{minsize}{Number indicating the minimum allowed size
  for the regulons, taken from \code{mobj} by default}

  \item{adaptive.size}{Logical, whether the weight
  (likelihood) should be used for computing the size, taken
  from \code{mobj} by default}

  \item{iterative}{Logical, whether a two step analysis
  with adaptive redundancy estimation should be performed,
  taken from \code{mobj} by default}

  \item{level}{Integer, maximum level of combinatorial
  regulation}

  \item{verbose}{Logical, whether progression messages
  should be printed in the terminal}
}
\value{
  A MARINA object
}
\description{
  This function performs combinatorial analysis for MARINA
  objects
}
\seealso{
  \code{\link{marina}}
}

