\name{marinaSynergy}
\alias{marinaSynergy}
\title{MARINa synergy analysis}
\usage{
  marinaSynergy(mobj, per = 1000, seed = 1, verbose = T)
}
\arguments{
  \item{mobj}{MARINa object containing combinatorial
  regulation results generated by
  \code{marinaCombinatorial}}

  \item{per}{Integer indicating the number of permutations}

  \item{seed}{Integer indicating the seed for the
  permutations, 0 for disable it}

  \item{verbose}{Logical, whether progression messages
  should be printed in the terminal}
}
\value{
  Updated MARINA object containing the sygergy p-value
}
\description{
  This function performs a synergy analysis for
  combinatorial regulation
}
\seealso{
  \code{\link{marina}}
}

