\name{pwea3NULLf}
\alias{pwea3NULLf}
\title{Null model function}
\usage{
  pwea3NULLf(pwnull, verbose = T)
}
\arguments{
  \item{pwnull}{Object generated by \code{pwea3NULLgroups}
  function}

  \item{verbose}{Logical, whether progression messages
  should be printed in the terminal}
}
\value{
  List of function to compute NES and p-value
}
\description{
  This function generates the NULL model function, which
  computes the normalized enrichment score and associated
  p-value
}

