\name{shadow}
\alias{shadow}
\title{Shadow analysis for MARINA objects}
\usage{
  shadow(mobj, pval = 0.01, targets = 10, shadow = 0.01,
    per = 1000, nullmodel = NULL, minsize = NULL,
    adaptive.size = NULL, iterative = NULL, seed = 1,
    verbose = T)
}
\arguments{
  \item{mobj}{MARINA object generated by \code{marina}}

  \item{pval}{Either a p-value cutoff, the total number of
  regulons to be used for computing the shadow effect, or a
  vector of TF ids to be considered}

  \item{targets}{Integer indicating the minimum number of
  common targets to compute shadow analysis}

  \item{shadow}{Number indicating the p-value threshold for
  the shadow effect}

  \item{per}{Integer indicating the number of permutations}

  \item{nullmodel}{Null model in marix format}

  \item{minsize}{Integer indicating the minimum size
  allowed for the regulons}

  \item{adaptive.size}{Logical, whether the target weight
  should be considered when computing the regulon size}

  \item{iterative}{Logical, whether a two step analysis
  with adaptive redundancy estimation should be performed}

  \item{seed}{Integer indicating the seed for the
  permutations, 0 for disable it}

  \item{verbose}{Logical, whether progression messages
  should be printed in the terminal}
}
\value{
  An updated MARINA object with an additional slot (shadow)
  containing the shadow pairs
}
\description{
  This function performs shadow analysis on MARINA objects
}
\seealso{
  \code{\link{marina}}
}

