\name{summary.marina}
\alias{summary.marina}
\title{List MARINA results}
\usage{
  \method{summary}{marina} (object, mrs = 10, ...)
}
\arguments{
  \item{object}{MARINA object}

  \item{mrs}{Either number of top MRs to report or vector
  containing the genes to display}

  \item{...}{Given for compatibility with the summary
  generic function}
}
\value{
  Data.frame with results
}
\description{
  This function generates a table of MARINA results
}

